+++ csatar-plugins/csatar/lang/en/lang.php
@@ -55,6 +55,7 @@
                 'bulkDeleteError' => 'The following error occurred during the delete process: ',
                 'activeMembersCount' => 'Active members count',
                 'clearFilters' => 'Clear filters',
+                'googleCalendarId' => 'Google Calendar ID',
             ],
             'history' => [
                 'histories' => 'History',
@@ -71,6 +72,9 @@
                 'newValue' => 'New value',
                 'description' => 'Description',
                 'ipAddress' => 'Ip address',
+            ],
+            'googleCalendarParams' => [
+                'googleCalendarParams' => 'Google Calendar Params for',
             ],
             'ageGroups' => [
                 'ageGroups' => 'Age Groups',
@@ -209,6 +213,10 @@
                 'activeMandateDeleteError' => 'The Scout having the %name name has active Mandates, thus this Scout cannot be deleted.',
                 'scoutTeam' => 'Scout\'s team',
                 'inactivationWarning' => 'Please note that if you change status from active to inactive, scout\'s all mandates will expire!',
+                'bulkDeletePartialSuccess' => ':deletedCount deleted from :totalCount. ',
+                'bulkDeleteError' => 'The following errors occurred: ',
+                'deleteWithPersonalData' => 'Delete with personal data',
+                'deleteWithPersonalDataConfirmation' => 'Are you sure you want to delete this scout and his personal data (name, email, phone, address?',
                 'teamChangeHistoryMessage' => 'On :date changed team from :oldTeam to :newTeam',
             ],
             'admin' => [