+++ Changelog.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - CS-533 - Modified search results to display results starting with Scouts to Association and last results from Content page
+- CS-555 - Added google calendar tab to Association, District and Team
+- CS-475 - Added option to backend scout list to (soft)delete scout with personal data
+- CS-520 - Modification and optimization of grid display and the design of the filtering
 - CS-491 - Added scout xlsx import/export to teams page
 - CS-331 - Added `manifest.json` file and updated readme with instructions on how to add PWA support
 - CS-549 - Fixed exception caused by missing getEagerLoadSettings method on AccidentLogRecord model